requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
mysqlclient==2.2.4
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
